🐰 feat: Gateの設定(admin)
- 投稿者本人と管理者のみが削除を可能にする

+++ app/Http/Controllers/PostController.php
@@ -86,7 +86,11 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
-        $post->delete();
-        return redirect()->route('post.index')->with('message', '投稿を削除しました');
+        if (Gate::any(['post-owner', 'admin'], $post)) {
+            $post->delete();
+            return redirect()->route('post.index')->with('message', '投稿を削除しました');
+        } else {
+            abort('403', 'Unauthorized action 🐶');
+        }
     }
 }
